LayoutTests/imported/w3c/web-platform-tests/html/browsers/browsing-the-web/history-traversal/unset_context_name-expected.txt


PASS Window.name is reset after navigating to a different origin
